chore(provider): merge server and storage packages
