+++ internal/service/storage/backup_rule.go
@@ -1,12 +1,7 @@
 package storage
 
 import (
-	"fmt"
-
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
-	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/request"
-	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/service"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -58,47 +53,4 @@ func BackupRule(backupRule map[string]interface{}) *upcloud.BackupRule {
 		}
 	}
 	return &upcloud.BackupRule{}
-}
-
-func ResizeStoragePartitionAndFs(client *service.Service, UUID, title string, deleteBackup bool) diag.Diagnostics {
-	diags := diag.Diagnostics{}
-
-	backup, err := client.ResizeStorageFilesystem(&request.ResizeStorageFilesystemRequest{
-		UUID: UUID,
-	})
-	if err != nil {
-		summary := fmt.Sprintf(
-			"Failed to resize partition and filesystem for storage %s(%s)",
-			UUID,
-			title,
-		)
-
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Warning,
-			Summary:  summary,
-			Detail:   err.Error(),
-		})
-	}
-
-	if err == nil && deleteBackup {
-		err = client.DeleteStorage(&request.DeleteStorageRequest{
-			UUID: backup.UUID,
-		})
-
-		if err != nil {
-			summary := fmt.Sprintf(
-				"Failed to delete the backup of storage %s(%s) after the partition and filesystem resize; you will need to delete the backup manually",
-				UUID,
-				title,
-			)
-
-			diags = append(diags, diag.Diagnostic{
-				Severity: diag.Warning,
-				Summary:  summary,
-				Detail:   err.Error(),
-			})
-		}
-	}
-
-	return diags
 }
