+++ internal/utils/server.go
@@ -1,0 +1,78 @@
+package utils
+
+import (
+	"log"
+	"time"
+
+	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
+	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/request"
+	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/service"
+)
+
+func VerifyServerStopped(stopRequest request.StopServerRequest, meta interface{}) error {
+	if stopRequest.Timeout == 0 {
+		stopRequest.Timeout = time.Minute * 2
+	}
+	if stopRequest.StopType == "" {
+		stopRequest.StopType = upcloud.StopTypeSoft
+	}
+
+	client := meta.(*service.Service)
+	// Get current server state
+	r := &request.GetServerDetailsRequest{
+		UUID: stopRequest.UUID,
+	}
+	server, err := client.GetServerDetails(r)
+	if err != nil {
+		return err
+	}
+	if server.State != upcloud.ServerStateStopped {
+		// Soft stop with 2 minute timeout, after which hard stop occurs
+		log.Printf("[INFO] Stopping server (server UUID: %s)", stopRequest.UUID)
+		_, err := client.StopServer(&stopRequest)
+		if err != nil {
+			return err
+		}
+		_, err = client.WaitForServerState(&request.WaitForServerStateRequest{
+			UUID:         stopRequest.UUID,
+			DesiredState: upcloud.ServerStateStopped,
+			Timeout:      time.Minute * 5,
+		})
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func VerifyServerStarted(startRequest request.StartServerRequest, meta interface{}) error {
+	if startRequest.Timeout == 0 {
+		startRequest.Timeout = time.Minute * 2
+	}
+
+	client := meta.(*service.Service)
+	// Get current server state
+	r := &request.GetServerDetailsRequest{
+		UUID: startRequest.UUID,
+	}
+	server, err := client.GetServerDetails(r)
+	if err != nil {
+		return err
+	}
+	if server.State != upcloud.ServerStateStarted {
+		log.Printf("[INFO] Starting server (server UUID: %s)", startRequest.UUID)
+		_, err := client.StartServer(&startRequest)
+		if err != nil {
+			return err
+		}
+		_, err = client.WaitForServerState(&request.WaitForServerStateRequest{
+			UUID:         startRequest.UUID,
+			DesiredState: upcloud.ServerStateStarted,
+			Timeout:      time.Minute * 5,
+		})
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
